Switched checkbox listener to the list adapter class
