+++ app/src/main/java/com/unigrade/app/View/Fragment/ClassesFragment.java
@@ -43,7 +43,10 @@ public class ClassesFragment extends Fragment {
     private TextView tvClassCredits;
     private TextView tvClassTitle;
     private String caller;
+
+
     private Subject subject;
+
     private ClassDAO classDAO;
     private SubjectDAO subjectDAO;
 
@@ -51,6 +54,9 @@ public class ClassesFragment extends Fragment {
         // Required empty public constructor
     }
 
+    public Subject getSubject() {
+        return subject;
+    }
 
     public void setClasses(ArrayList<SubjectClass> classes){
         this.classes = classes;
@@ -104,7 +110,6 @@ public class ClassesFragment extends Fragment {
         }
 
         classesList.setItemsCanFocus(false);
-        classesList.setOnItemClickListener(getItemListener());
 
         btnReload.setOnClickListener(new View.OnClickListener() {
             @Override
@@ -137,68 +142,11 @@ public class ClassesFragment extends Fragment {
         for (SubjectClass sc: classes)
             Log.i("ISSELECTED", String.valueOf(sc.isSelected()));
         classesList.setAdapter(
-                new ClassListAdapter(classes, getActivity())
+                new ClassListAdapter(classes, getActivity(), subject)
         );
     }
 
-    private AdapterView.OnItemClickListener getItemListener(){
-        return new AdapterView.OnItemClickListener() {
-            @Override
-            public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
-                CheckBox cb = view.findViewById(R.id.checkbox);
 
-                for (SubjectClass c : classes)
-                    c.setSubjectCode(subject.getCode());
-
-                SubjectClass sc = (SubjectClass) parent.getItemAtPosition(position);
-
-                if (!cb.isChecked()) {
-                    cb.setChecked(true);
-                    sc.setSelected(true);
-                    insertIntoDatabase(sc);
-                    Log.i("ADDED", sc.getTeacher() + " " + sc.isSelected());
-                } else {
-                    cb.setChecked(false);
-                    sc.setSelected(false);
-                    removeFromDatabase(sc);
-                    Log.i("REMOVED", sc.getTeacher() + " " + sc.isSelected());
-                }
-
-            }
-        };
-    }
-
-    private void insertIntoDatabase(SubjectClass sc){
-        if (!subjectDAO.isSubjectOnDB(subject.getCode())){
-            subjectDAO.insert(subject);
-            for (SubjectClass c : classes)
-                classDAO.insert(c);
-            Log.i("OUTSIDEDB", subject.getCode() + " "+ sc.getTeacher());
-        } else {
-            classDAO.alter(sc);
-            Log.i("ONDB", subject.getCode() + " "+ sc.getTeacher());
-        }
-    }
-
-    private void removeFromDatabase(SubjectClass sc){
-        if (isLonelyAdded(sc)){
-            for (SubjectClass c : classes)
-                classDAO.delete(c);
-            subjectDAO.delete(subject);
-            Log.i("LONELY", subject.getCode() + " "+ sc.getTeacher());
-        } else {
-            classDAO.alter(sc);
-            Log.i("NOTLONELY", subject.getCode() + " "+ sc.getTeacher());
-        }
-    }
-
-    private boolean isLonelyAdded(SubjectClass sc){
-        for (SubjectClass c : classes)
-            if (c.isSelected() && c != sc)
-                return false;
-
-        return true;
-    }
 
     public interface OnFragmentInteractionListener {
         // TODO: Update argument type and name
